refactor(content): utils import src change
